+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.0.5
 ======
 
 * Add an implementation of Tarjan's algorithm (alg.tarjan).
+* Add a function to find and return all cycles in a graph (alg.findCycles).
 
 v0.0.4
 ======